tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/events/001.html

<!doctype html>
<title>WebSockets: getting on*</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<meta name="variant" content="">
<meta name="variant" content="?wss">
<div id=log></div>
<script>
var events = ['open', 'message', 'error', 'close'];
for (var i = 0; i < events.length; ++i) {
  test(function(){
    var ws = new WebSocket(SCHEME_DOMAIN_PORT+'/');
    assert_equals(ws['on'+events[i]], null, 'on'+events[i]);
  });
};
</script>
